define ValueSet#substract and ValueSet#add to help switching to Set
ValueSet should really go away. The main interest of the class was to
avoid garbage collections, and since 2.1+ it has become part of the
problem (since it is not compatible with Ruby's generational GC)

This defines two methods that are implemented in ValueSet and Set but
have a different name in both, so that ValueSet-using code can switch
to them.
